server/requirements.txt
flask
flask_sqlalchemy
mysqlclient
flask_cors
flask_jwt